# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lensio API
  version: 1.0.0
  description: |
    Indonesian KTP OCR API as a Service.
    Affordable, privacy-conscious, and production-grade identity data extraction.
  contact:
    name: Lensio Engineering Team
    url: https://lensio.dev
    email: api@lensio.dev
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.staging.lensio.dev
    description: Staging environment
  - url: https://api.lensio.dev
    description: Production environment
...
paths:
...
  /api/v1/auth/api-keys:
    post:
      summary: Create API Key
      description: |
        Generates a new cryptographically secure API key.
        The plaintext token is displayed only once upon creation and cannot be retrieved later.
      operationId: createAPIKey
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Human-readable label for the API key
                  example: Production Backend Service
                environment:
                  type: string
                  enum: [live, test]
                  default: live
                  example: live
                scopes:
                  type: array
                  items:
                    type: string
                  example: ["ocr:read", "ocr:write", "usage:read"]
                expires_at:
                  type: string
                  format: date-time
                  nullable: true
                  description: Optional expiry timestamp
                  example: "2027-01-01T00:00:00Z"
                org_id:
                  type: string
                  format: uuid
                  description: Organization ID (optional if authenticated or default org used)
      responses:
        '201':
          description: API key created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  org_id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  key:
                    type: string
                    description: Plaintext API key token (returned only once)
                    example: lensio_live_a1b2c3d4e5f6...
                  prefix:
                    type: string
                    example: lensio_live_a1b2
                  scopes:
                    type: array
                    items:
                      type: string
                  environment:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/InvalidAPIKey'

    get:
      summary: List API Keys
      description: Lists active and revoked API keys for the caller's organization. Secrets are masked.
      operationId: listAPIKeys
      tags:
        - Authentication
      security:
        - BearerAuth: []
        - {}
      responses:
        '200':
          description: List of organization API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKeyListItem'
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
  /api/v1/auth/api-keys/{id}:
    delete:
      summary: Revoke API Key
      description: Immediately revokes an API key, preventing any further requests using it.
      operationId: revokeAPIKey
      tags:
        - Authentication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: API key successfully revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: API key successfully revoked
                  id:
                    type: string
                    format: uuid
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/usage:
    get:
      summary: Get Usage Summary
      description: Retrieves organization usage metrics and quota consumption for current billing cycle.
      operationId: getUsageSummary
      tags:
        - Usage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Organization usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
        '403':
          $ref: '#/components/responses/InsufficientScope'
  /api/v1/usage/daily:
    get:
      summary: Get Daily Usage Timeseries
      description: Retrieves daily breakdown of requests and status codes.
      operationId: getDailyUsage
      tags:
        - Usage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Daily usage timeseries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        total_requests:
                          type: integer
                        success_count:
                          type: integer
                        error_count:
                          type: integer
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
  /api/v1/usage/endpoints:
    get:
      summary: Get Usage by Endpoint
      description: Retrieves breakdown of requests partitioned by endpoint.
      operationId: getEndpointUsage
      tags:
        - Usage
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Endpoint usage breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        endpoint:
                          type: string
                        total_requests:
                          type: integer
                        avg_latency_ms:
                          type: number
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
  /api/v1/usage/records:
    get:
      summary: Get Request Execution Logs
      description: Retrieves paginated request logs with latency and status codes for the Request Explorer (strictly non-PII).
      operationId: getUsageRecords
      tags:
        - Usage
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: status_code
          in: query
          required: false
          schema:
            type: integer
        - name: endpoint
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated request log records
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        request_id:
                          type: string
                        endpoint:
                          type: string
                        status_code:
                          type: integer
                        latency_ms:
                          type: integer
                        timestamp:
                          type: string
                          format: date-time
        '401':
          $ref: '#/components/responses/InvalidAPIKey'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Lensio API Key (lensio_live_... or lensio_test_...)
...
  schemas:
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_request
                - invalid_api_key
                - insufficient_scope
                - rate_limit_exceeded
                - quota_exceeded
                - invalid_document
                - unsupported_document
                - ocr_failed
                - low_confidence
                - internal_error
              example: invalid_document
            message:
              type: string
              example: The uploaded document could not be processed
            request_id:
              type: string
              example: req_01JABC12345
    APIKeyListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        prefix:
          type: string
          example: lensio_live_a1b2
        masked_key:
          type: string
          example: lensio_live_a1b2••••••••
        scopes:
          type: array
          items:
            type: string
        environment:
          type: string
          example: live
        last_used_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    UsageSummary:
      type: object
      properties:
        total_requests:
          type: integer
          example: 450
        quota_limit:
          type: integer
          example: 1000
        quota_remaining:
          type: integer
          example: 550
        billing_cycle_reset:
          type: string
          format: date-time
  responses:
    InvalidRequest:
      description: Malformed request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: invalid_request
              message: Missing required field 'name'
              request_id: req_01JABC12345
    InvalidAPIKey:
      description: Authentication failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: invalid_api_key
              message: Invalid or unrecognized API key
              request_id: req_01JABC12345
    InsufficientScope:
      description: Authorization failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: insufficient_scope
              message: "API key lacks required scope: ocr:write"
              request_id: req_01JABC12345
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: invalid_request
              message: Resource not found
              request_id: req_01JABC12345